+++ test/assets/module.js
@@ -1,0 +1,2 @@
+pathsLoaded['module.js'] = (pathsLoaded['module.js'] || 0) + 1;
+