+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # JohnnyDep Changelog
 
+## 0.1.4 - August 27, 2021
+
+* Bumped path-parse from 1.0.6 to 1.0.7
+
+## 0.1.3 - April 22, 2021
+
+* Added support for JavaScript modules with "module!" modifier
+
+## 0.1.2 - April 1, 2021
+
+* Upgraded devDependencies
+
+## 0.1.1 - December 11, 2020
+
+* Upgraded devDependencies
+
+## 0.1.0 - March 19, 2019
+
+* Fixed bug in IE Edge causing CSS load failures to call success callbacks
+
+## 0.0.3 - October 15, 2018
+
+* Added .done() method
+* Added support for defining bundles after calling require
+
 ## 0.0.2 - September 28, 2018
 
 * Added missing files